include file name in error handling
