+++ migration/src/transformCliUtils.ts
@@ -6,12 +6,16 @@ const transformCliUtils = (node: ts.Node, importName: string) => {
   const isUtilCall = isCallExpression(node) &&
     ts.isPropertyAccessExpression(node.expression) &&
     ts.isIdentifier(node.expression.expression) &&
-    node.expression.expression.escapedText === 'cli'
+    node.expression.expression.escapedText.toString() === importName
+
   if (!isUtilCall) {
     return node
   }
 
+  const callName = node.expression.name.escapedText.toString()
+
   // transform
+  debugger
 }
 
 export default transformCliUtils
